Automatically enable new cops
When new cops are added to Rubocop, they are by default in pending
status which means that the user has to explicitly enable or disable
them. This feature was added so that when Rubocop is updated, existing
integrations would not break.

As most of the new cops are useful, I think it makes sense to enable
them by default. By setting the `NewCops` flag to `enable`, all cops
in pending status will be enabled.

CHAN-324