Update rubocop to the latest version
It's good to be up to date.

One new cop `RSpec/NoExpectationExample` is disabled because it gave
us quite a few false positives. Namely, this cop requires that in each
test you have a `expect` call. In many tests in our code the `expect`
function is not called directly in the test, but rather in some helper
function.