Disable RSpec/ExampleLength
We already have it disabled in many repos and I think it makes sense.
Frequently tests can get quite long due to setup and the fact that
some things are not easy to test. By spreading the code in `before`
blocks and contexts usually leads to worser readability than having
long test examples.